Check if the query results exists

+++ program/lib/Roundcube/rcube_session.php
@@ -409,11 +409,13 @@ class rcube_session
                 $res = $this->db->query("SELECT sess_id, created, changed, ip FROM " . $this->db->table_name('session')
                     . " WHERE changed < " . $this->db->now(-$this->gc_enabled));
 
-                while ($sql_arr = $this->db->fetch_assoc($res)) {
-                    $data = rcmail::get_instance()->plugins->exec_hook('session_garbage', array('session' => $sql_arr));
-                    if (!$data['abort']) {
-                        // delete the expired session
-                        $this->db->query("DELETE FROM " . $this->db->table_name('session') . " WHERE sess_id = ?", $sql_arr['sess_id']);
+                if ($res) {
+                    while ($sql_arr = $this->db->fetch_assoc($res)) {
+                        $data = rcmail::get_instance()->plugins->exec_hook('session_garbage', array('session' => $sql_arr));
+                        if (!$data['abort']) {
+                            // delete the expired session
+                            $this->db->query("DELETE FROM " . $this->db->table_name('session') . " WHERE sess_id = ?", $sql_arr['sess_id']);
+                        }
                     }
                 }
             }
